support for <html5> datalist
git-svn-id: http://sources.disruptive-innovations.com/bluegriffon/trunk@340 c25cacc5-b657-0410-8d04-e2f7359d2943

+++ base/content/bluegriffon/dialogs/insertDatalist.js
@@ -11,7 +11,6 @@ function Startup()
   GetUIElements();
 
   if (gNode) {
-    InitDialog();
     InitDialog2(gNode, gDialog.contentsTree);
   }
 }
@@ -22,16 +21,15 @@ function onAccept()
 
   var doc = EditorUtils.getCurrentDocument();
   if (!gNode) {
-    gNode = doc.createElement("select");
+    gNode = doc.createElement("datalist");
     gEditor.insertElementAtSelection(gNode, true);
   }
 
-  ApplyAttributes();
-
-  var child = gNode.lastChild;
+  var child = gNode.lastElementChild;
   while (child) {
-    var tmp = child.previousSibling;
-    gEditor.deleteNode(child);
+    var tmp = child.previousElementSibling;
+    if (tmp.nodename.toLowerCase() == "option")
+      gEditor.deleteNode(child);
     child = tmp;
   }
 
@@ -40,26 +38,14 @@ function onAccept()
   for (var i = 0; i < treeitems.length; i++)
   {
     var treeitem = treeitems[i];
-    if (treeitem.hasAttribute("container")) {
-      var child = doc.createElement("optgroup");
-      child.setAttribute("label", treeitem.firstChild.childNodes[1].getAttribute("label"));
-      if (treeitem.firstChild.childNodes[2].getAttribute("label") == "✔")
-        child.setAttribute("disabled", "disabled");
-      parent = child;
-      gEditor.insertNode(child, gNode, gNode.childNodes.length);
-    }
-    else {
-      if (treeitem.parentNode.parentNode.id == "contentsTree")
-        parent = gNode;
-      var child = doc.createElement("option");
-      child.setAttribute("name", treeitem.firstChild.childNodes[0].getAttribute("label"));
-      child.textContent = treeitem.firstChild.childNodes[1].getAttribute("label");
-      if (treeitem.firstChild.childNodes[2].getAttribute("label") == "✔")
-        child.setAttribute("disabled", "disabled");
-      if (treeitem.firstChild.childNodes[3].getAttribute("label"))
-      child.setAttribute("selected", "selected");
-      gEditor.insertNode(child, parent, parent.childNodes.length);
-    }
+    var child = doc.createElement("option");
+    child.setAttribute("name", treeitem.firstChild.childNodes[0].getAttribute("label"));
+    child.textContent = treeitem.firstChild.childNodes[1].getAttribute("label");
+    if (treeitem.firstChild.childNodes[2].getAttribute("label") == "✔")
+      child.setAttribute("disabled", "disabled");
+    if (treeitem.firstChild.childNodes[3].getAttribute("label"))
+    child.setAttribute("selected", "selected");
+    gEditor.insertNode(child, gNode, gNode.childNodes.length);
   }
 
   gEditor.endTransaction();
@@ -149,18 +135,6 @@ function UpdateButtons()
   gDialog.DownButton.disabled = !treeitem.nextElementSibling;
 }
 
-function AddOptgroup(aWhere)
-{
-  gWhere = aWhere;
-  gDialog.doUpdateOptGroupButton.hidden = true;
-  gDialog.doAddOptGroupButton.hidden = false;
-  gDialog.optGroupLabelTextbox.value = "";
-  gDialog.optGroupDisabledCheckbox.checked = false;
-  gDialog.optGroupPanel.openPopup(gDialog.PlusButton,
-                                 "after_start", 0, 0,
-                                 false, true);
-}
-
 function DeleteOpt()
 {
   var tree = gDialog.contentsTree;
@@ -175,54 +149,11 @@ function DeleteOpt()
   UpdateButtons();
 }
 
-function doAddOptGroup()
+function AddOption()
 {
-  var tree = gDialog.contentsTree;
-  var contentView = tree.contentView;
-  var view = tree.view;
-  if (!view || !view.selection) { // nothing in the tree
-    item = AddTreeItem(tree);
-  }
-  else {
-	  var index = view.selection.currentIndex;
-	  var treeitem = contentView.getItemAtIndex(index);
-	
-    if (gWhere == "before" || gWhere == "after") {
-	    item = document.createElement("treeitem");
-	    var row = document.createElement("treerow");
-	    item.appendChild(row);
-	    if (gWhere == "before")
-		    treeitem.parentNode.insertBefore(item, treeitem);
-	    else
-	      treeitem.parentNode.insertBefore(item, treeitem.nextSibling);
-    }
-    else if (gWhere == "inside")
-      item = AddTreeItem(treeitem);
-  }
-
-  var cell1 = document.createElement("treecell");
-	cell1.setAttribute("label", "");
-	var cell2 = document.createElement("treecell");
-	cell2.setAttribute("label", gDialog.optGroupLabelTextbox.value);
-	var cell3 = document.createElement("treecell");
-	cell3.setAttribute("label", gDialog.optGroupDisabledCheckbox.checked ? "✔" : "");
-	var cell4 = document.createElement("treecell");
-	cell4.setAttribute("label", "");
-	item.firstChild.appendChild(cell1);
-	item.firstChild.appendChild(cell2);
-	item.firstChild.appendChild(cell3);
-	item.firstChild.appendChild(cell4);
-  item.setAttribute("container", "true");
-  gDialog.optGroupPanel.hidePopup();
-  gDialog.contentsTree.view.selection.select(gDialog.contentsTree.contentView.getIndexOfItem(item));
-  UpdateButtons();
-}
-
-function AddOption(aWhere)
-{
-  gWhere = aWhere;
   gDialog.doUpdateOptionButton.hidden = true;
   gDialog.doAddOptionButton.hidden = false;
+
   gDialog.optionLabelTextbox.value = "";
   gDialog.optionValueTextbox.value = "";
   gDialog.optionSelectedCheckbox.checked = false;
@@ -237,25 +168,8 @@ function doAddOption()
   var tree = gDialog.contentsTree;
   var contentView = tree.contentView;
   var view = tree.view;
-  if (!view || !view.selection) { // nothing in the tree
-    item = AddTreeItem(tree);
-  }
-  else {
-    var index = view.selection.currentIndex;
-    var treeitem = contentView.getItemAtIndex(index);
-  
-    if (gWhere == "before" || gWhere == "after") {
-      item = document.createElement("treeitem");
-      var row = document.createElement("treerow");
-      item.appendChild(row);
-      if (gWhere == "before")
-        treeitem.parentNode.insertBefore(item, treeitem);
-      else
-        treeitem.parentNode.insertBefore(item, treeitem.nextSibling);
-    }
-    else if (gWhere == "inside")
-      item = AddTreeItem(treeitem);
-  }
+
+  var item = AddTreeItem(tree);
 
   var cell1 = document.createElement("treecell");
   cell1.setAttribute("label", gDialog.optionLabelTextbox.value);
@@ -274,33 +188,6 @@ function doAddOption()
   gDialog.optionPanel.hidePopup();
   gDialog.contentsTree.view.selection.select(gDialog.contentsTree.contentView.getIndexOfItem(item));
   UpdateButtons();
-}
-
-function UpdateOptions()
-{
-  var tree = gDialog.contentsTree;
-  var contentView = tree.contentView;
-  var view = tree.view;
-  if (!view || !view.selection) { // no selection...
-    //gDialog.insideMenu.disabled = true;
-    gDialog.insideMenu.disabled = true;
-    gDialog.addOptionInsideMenuitem.disabled = true;
-    gDialog.addOptgroupInsideMenuitem.disabled = true;
-    gDialog.beforeMenu.disabled = true;
-    gDialog.afterMenu.disabled = false;
-    return;
-  }
-
-  gDialog.beforeMenu.disabled = false;
-  var index = view.selection.currentIndex;
-  var treeitem = contentView.getItemAtIndex(index);
-  var isTopLevel = (treeitem.parentNode.parentNode.id == "contentsTree");
-  gDialog.addOptgroupBeforeMenuitem.disabled = !isTopLevel;
-  gDialog.addOptgroupAfterMenuitem.disabled = !isTopLevel;
-
-  gDialog.insideMenu.disabled = false;
-  gDialog.addOptionInsideMenuitem.disabled = !treeitem.hasAttribute("container");
-  gDialog.addOptGroupInsideMenuitem.disabled = true;
 }
 
 function Down()
@@ -331,14 +218,6 @@ function Up()
   UpdateButtons();
 }
 
-function CheckCRInOptGroupPanel(aEvent)
-{
-  if (aEvent.keyCode == 13) { // CR key
-    aEvent.preventDefault();
-    doAddOptGroup();
-  }
-}
-
 function CheckCRInOptionPanel(aEvent)
 {
   if (aEvent.keyCode == 13) { // CR key
@@ -354,56 +233,16 @@ function UpdateUrl()
   var view = tree.view;
   var index = view.selection.currentIndex;
   var treeitem = contentView.getItemAtIndex(index);
-  var isContainer = treeitem.hasAttribute("container");
-  if (isContainer) {
-    gDialog.doUpdateOptGroupButton.hidden = false;
-    gDialog.doAddOptGroupButton.hidden = true;
-	  gDialog.optGroupLabelTextbox.value = treeitem.firstChild.childNodes[1].getAttribute("label");
-	  gDialog.optGroupDisabledCheckbox.checked = (treeitem.firstChild.childNodes[2].getAttribute("label") == "✔");
-	  gDialog.optGroupPanel.openPopup(gDialog.ConfigButton,
-	                                  "after_start", 0, 0,
-	                                  false, true);
-  }
-  else {
-	  gDialog.doUpdateOptionButton.hidden = false;
-	  gDialog.doAddOptionButton.hidden = true;
-	  gDialog.optionLabelTextbox.value = treeitem.firstChild.childNodes[0].getAttribute("label");
-	  gDialog.optionValueTextbox.value = treeitem.firstChild.childNodes[1].getAttribute("label");
-	  gDialog.optionSelectedCheckbox.checked = (treeitem.firstChild.childNodes[2].getAttribute("label") == "✔");;
-	  gDialog.optionDisabledCheckbox.checked = (treeitem.firstChild.childNodes[3].getAttribute("label") == "✔");;
-	  gDialog.optionPanel.openPopup(gDialog.ConfigButton,
-	                                 "after_start", 0, 0,
-	                                 false, true);
-  }
-}
 
-function doUpdateOptGroup()
-{
-  var tree = gDialog.contentsTree;
-  var contentView = tree.contentView;
-  var view = tree.view;
-  var index = view.selection.currentIndex;
-  var item = contentView.getItemAtIndex(index);
-  item.removeChild(item.firstChild);
-  var treerow = document.createElement("treerow");
-  item.appendChild(treerow);
-
-  var cell1 = document.createElement("treecell");
-  cell1.setAttribute("label", "");
-  var cell2 = document.createElement("treecell");
-  cell2.setAttribute("label", gDialog.optGroupLabelTextbox.value);
-  var cell3 = document.createElement("treecell");
-  cell3.setAttribute("label", gDialog.optGroupDisabledCheckbox.checked ? "✔" : "");
-  var cell4 = document.createElement("treecell");
-  cell4.setAttribute("label", "");
-  item.firstChild.appendChild(cell1);
-  item.firstChild.appendChild(cell2);
-  item.firstChild.appendChild(cell3);
-  item.firstChild.appendChild(cell4);
-  item.setAttribute("container", "true");
-  gDialog.optGroupPanel.hidePopup();
-  gDialog.contentsTree.view.selection.select(gDialog.contentsTree.contentView.getIndexOfItem(item));
-  UpdateButtons();
+  gDialog.doUpdateOptionButton.hidden = false;
+  gDialog.doAddOptionButton.hidden = true;
+  gDialog.optionLabelTextbox.value = treeitem.firstChild.childNodes[0].getAttribute("label");
+  gDialog.optionValueTextbox.value = treeitem.firstChild.childNodes[1].getAttribute("label");
+  gDialog.optionSelectedCheckbox.checked = (treeitem.firstChild.childNodes[2].getAttribute("label") == "✔");;
+  gDialog.optionDisabledCheckbox.checked = (treeitem.firstChild.childNodes[3].getAttribute("label") == "✔");;
+  gDialog.optionPanel.openPopup(gDialog.ConfigButton,
+                                 "after_start", 0, 0,
+                                 false, true);
 }
 
 function doUpdateOption()
@@ -429,7 +268,6 @@ function doUpdateOption()
   item.firstChild.appendChild(cell2);
   item.firstChild.appendChild(cell3);
   item.firstChild.appendChild(cell4);
-
 
   gDialog.optionPanel.hidePopup();
   gDialog.contentsTree.view.selection.select(gDialog.contentsTree.contentView.getIndexOfItem(item));
